requirements.txt
environs>=9.5.0
asgiref>=3.6.0
Django>=4.2.1
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.2
PyJWT>=2.6.0
pytz>=2023.3
sqlparse>=0.4.4
mysqlclient>=2.1.1
pillow>=9.5.0
drf-yasg>=1.21.5
